// Downloads/dpm-ui-web-and-mobile-fixes-21-06-2025/dpm-ui-web-and-mobile-fixes/web-apps/consumer-portal/src/types/coverageplan.ts
export interface IDetails {
  id: number;
  itemname: string;
}

export type ICoveragePlanData = {
  title: string;
  key: string;
  details: IDetails[];
  image?: string;
};

export type CompensationTypeKeys = "comprehensive" | "thirdparty" | "buildingcontents" | "contents";

export type SelectPlanTypeKeys = "worldwide" | "worldwide_except_usa" | "europe";

export const IWorldwide: string = "worldwide";
export const IworldwideExcept: string = "worldwide_except_usa";
export const IEuropedata:string= "europe"